split models into separate file from db utils
Moves all base model objects from resource_models.py into models.py and
renames resource_models.py to db.py.
