+++ lookup.py
@@ -1,13 +1,11 @@
 # Simple functions for getting heavily-cached simple lookups for type/class
 # information
 
-import os
-import subprocess
-
 import sqlalchemy as sa
 
+import db
 import metadata
-import resource_models
+import models
 
 _PROVIDER_TYPE_MAP = None
 _CONSUMER_TYPE_MAP = None
@@ -20,9 +18,9 @@ def _get_provider_type_map():
     global _PROVIDER_TYPE_MAP
     if _PROVIDER_TYPE_MAP is not None:
         return _PROVIDER_TYPE_MAP
-    tbl = resource_models.get_table('provider_types')
+    tbl = db.get_table('provider_types')
     sel = sa.select([tbl.c.id, tbl.c.code])
-    sess = resource_models.get_session()
+    sess = db.get_session()
     _PROVIDER_TYPE_MAP = {r[1]: r[0] for r in sess.execute(sel)}
     return _PROVIDER_TYPE_MAP
 
@@ -38,9 +36,9 @@ def _get_consumer_type_map():
     global _CONSUMER_TYPE_MAP
     if _CONSUMER_TYPE_MAP is not None:
         return _CONSUMER_TYPE_MAP
-    tbl = resource_models.get_table('consumer_types')
+    tbl = db.get_table('consumer_types')
     sel = sa.select([tbl.c.id, tbl.c.code])
-    sess = resource_models.get_session()
+    sess = db.get_session()
     _CONSUMER_TYPE_MAP = {r[1]: r[0] for r in sess.execute(sel)}
     return _CONSUMER_TYPE_MAP
 
@@ -56,9 +54,9 @@ def _get_resource_type_map():
     global _RESOURCE_TYPE_MAP
     if _RESOURCE_TYPE_MAP is not None:
         return _RESOURCE_TYPE_MAP
-    tbl = resource_models.get_table('resource_types')
+    tbl = db.get_table('resource_types')
     sel = sa.select([tbl.c.id, tbl.c.code])
-    sess = resource_models.get_session()
+    sess = db.get_session()
     _RESOURCE_TYPE_MAP = {r[1]: r[0] for r in sess.execute(sel)}
     return _RESOURCE_TYPE_MAP
